delete deprecated gorgias api key app
