--- components/gorgias/README.md
@@ -1,14 +1,0 @@
-# Overview
-
-The Gorgias API lets you build powerful email automation tools. With it, you
-can:
-
-- Automatically send customized emails to customers based on their interactions
-  with your business
-- Trigger emails based on customer behavior, such as abandoned cart emails or
-  order follow-ups
-- Manage your email list and subscribers
-- Analyze your email campaigns to see what's working and what's not
-
-Check out the Gorgias API documentation to learn more about what you can build
-with it.
