--- components/gorgias/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@pipedream/gorgias",
-  "version": "0.3.5",
-  "description": "Pipedream Gorgias Components",
-  "main": "gorgias.app.mjs",
-  "keywords": [
-    "pipedream",
-    "gorgias"
-  ],
-  "homepage": "https://pipedream.com/apps/gorgias",
-  "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
-  "gitHead": "e12480b94cc03bed4808ebc6b13e7fdb3a1ba535",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@pipedream/platform": "^0.10.0",
-    "lodash-es": "^4.17.21"
-  }
-}
