+++ components/gorgias_oauth/actions/retrieve-customer/retrieve-customer.mjs
@@ -1,14 +1,26 @@
-import base from "../../../gorgias/actions/retrieve-customer/retrieve-customer.mjs";
-import overrideApp from "../../common/override-app.mjs";
+import gorgias from "../../gorgias_oauth.app.mjs";
 
-overrideApp(base);
-
-// same version as base
-// eslint-disable-next-line pipedream/required-properties-version
 export default {
-  ...base,
   key: "gorgias_oauth-retrieve-customer",
   name: "Retrieve a Customer",
   description: "Retrieve a customer. [See the docs](https://developers.gorgias.com/reference/get_api-customers-id-)",
+  version: "0.0.1",
   type: "action",
+  props: {
+    gorgias,
+    customerId: {
+      propDefinition: [
+        gorgias,
+        "customerId",
+      ],
+    },
+  },
+  async run({ $ }) {
+    const response = await this.gorgias.retrieveCustomer({
+      $,
+      id: this.customerId,
+    });
+    $.export("$summary", `Succesfully retrieved customer ${this.customerId}`);
+    return response;
+  },
 };
