+++ components/gorgias_oauth/common/languages.mjs
