+++ components/gorgias_oauth/package.json
@@ -15,6 +15,7 @@
     "access": "public"
   },
   "dependencies": {
-    "@pipedream/platform": "^0.10.0"
+    "@pipedream/platform": "^1.4.1",
+    "lodash-es": "^4.17.21"
   }
 }
