+++ components/gorgias_oauth/sources/ticket-message-created/ticket-message-created.mjs
@@ -1,13 +1,26 @@
-import base from "../../../gorgias/sources/ticket-message-created/ticket-message-created.mjs";
-import overrideApp from "../../common/override-app.mjs";
+import base from "../common/base.mjs";
+import eventTypes from "../common/event-types.mjs";
 
-overrideApp(base);
-
-// eslint-disable-next-line pipedream/required-properties-version
 export default {
   ...base,
   key: "gorgias_oauth-ticket-messaged-created",
   name: "New Ticket Message",
   description: "Emit new event when a ticket message is created. [See the docs](https://developers.gorgias.com/reference/the-event-object)",
+  version: "0.1.1",
   type: "source",
+  methods: {
+    ...base.methods,
+    getEventType() {
+      return eventTypes.TICKET_MESSAGE_CREATED;
+    },
+    async processHistoricalEvent(event) {
+      // event doesn't contain ticket id to fetch message
+      return {
+        event,
+      };
+    },
+    async processEvent(event) {
+      this.emitEvent(event.message);
+    },
+  },
 };
