+++ components/gorgias_oauth/sources/ticket-updated/ticket-updated.mjs
@@ -1,13 +1,26 @@
-import base from "../../../gorgias/sources/ticket-updated/ticket-updated.mjs";
-import overrideApp from "../../common/override-app.mjs";
+import base from "../common/base.mjs";
+import eventTypes from "../common/event-types.mjs";
 
-overrideApp(base);
-
-// eslint-disable-next-line pipedream/required-properties-version
 export default {
   ...base,
   key: "gorgias_oauth-ticket-updated",
   name: "New Updated Ticket",
   description: "Emit new event when a ticket is updated. [See the docs](https://developers.gorgias.com/reference/the-event-object)",
+  version: "0.1.1",
   type: "source",
+  methods: {
+    ...base.methods,
+    getEventType() {
+      return eventTypes.TICKET_UPDATED;
+    },
+    async processHistoricalEvent(event) {
+      const ticket = await this.retrieveTicket(event.object_id);
+      return {
+        ticket,
+      };
+    },
+    async processEvent(event) {
+      this.emitEvent(event.ticket);
+    },
+  },
 };
